requirements.txt
anyascii==0.3.0
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==3.2.10
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-modelcluster==5.2
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
l18n==2021.3
openpyxl==3.0.9
Pillow==8.4.0
psycopg2==2.9.3
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.1.0
telepath==0.2
urllib3==1.26.7
wagtail==2.15.1
webencodings==0.5.1
whitenoise==5.3.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
